+++ src/utils/loadComments.ts
@@ -1,16 +1,17 @@
 import * as vscode from "vscode";
 
+import { splitUri } from "./splitUri";
 import { getFileDecorator } from "../config/globals";
-import { getRootFolderPath } from "../utils/reviewRootDir";
 
 export async function loadFileDecorator() {
   const editor = vscode.window.activeTextEditor;
   if (!editor) {
     return;
   }
+
   const doc = editor.document;
-  const fullPath = doc.uri.fsPath;
-  const rootFolder = await getRootFolderPath();
+  const {fullPath, rootFolder} = await splitUri(doc.uri);
+
   if (!fullPath.startsWith(rootFolder)) {
     return;
   }
