+++ src/utils/storage.ts
@@ -1,25 +1,26 @@
 import * as vscode from "vscode";
 
+import { addToCache, getFromCache } from "./addonCache";
+import { createComment } from "./comment";
+import getCommentLocation, {  stringToRange } from "./getCommentLocation";
 import { loadFileDecorator } from "./loadComments";
+import { splitUri } from "./splitUri";
 import { AssayComment } from "../config/comment";
-import { addToCache, getFromCache } from "../utils/addonCache";
-import createComment from '../utils/createComment';
-import getCommentLocation, { splitUri, stringToRange } from "../utils/getCommentLocation";
 
 export async function fetchCommentsFromCache(controller: vscode.CommentController){
+  const editor = vscode.window.activeTextEditor;
+  if (!editor) {
+      return;
+  }
 
-    const editor = vscode.window.activeTextEditor;
-    if (!editor) {
-        return;
-    }
-    const doc = editor.document;
-    
-    const {rootFolder, fullPath, guid} = await splitUri(doc.uri);
-    if (!fullPath.startsWith(rootFolder)) {
-        return;
-    }
+  const doc = editor.document;
+  const {rootFolder, fullPath, guid} = await splitUri(doc.uri);
+  
+  if (!fullPath.startsWith(rootFolder)) {
+      return;
+  }
 
-    const comments = await getFromCache(guid, ['comments']);
+  const comments = await getFromCache(guid, ['comments']);
 
   for(const version in comments){
     for (const filepath in comments[version]) {
@@ -31,7 +32,6 @@ export async function fetchCommentsFromCache(controller: vscode.CommentControlle
         }
       }
   }
-
 }
 
 export async function saveCommentToCache(comment: AssayComment){
